build: Migrate from tsdx to Rollup for modern build tooling

+++ scripts/organize-styles.sh
@@ -1,7 +1,16 @@
 #!/usr/bin/env sh
 
-mv es/index.css es/reactist.css
+# Rename the main index CSS file to reactist.css
+[ -f es/index.css ] && mv es/index.css es/reactist.css
+
+# Create styles directory
 mkdir -p styles
-find es -iname '*.css' -type f -exec mv {} styles/ \;
-find dist -iname '*.css' -type f -exec rm {} \;
-find lib -iname '*.css' -type f -exec rm {} \;
+
+# Move all CSS files from es/ to styles/
+[ -d es ] && find es -iname '*.css' -type f -exec mv {} styles/ \;
+
+# Delete CSS files from dist/ and lib/ (we only keep them in styles/)
+[ -d dist ] && find dist -iname '*.css' -type f -exec rm {} \;
+[ -d lib ] && find lib -iname '*.css' -type f -exec rm {} \;
+
+printf "\n💅 Styles reorganization finished successfully.\n\n"
